rename data to structures (#1854)
