+++ mmseg/models/segmentors/base.py
@@ -6,8 +6,8 @@ from mmengine.data import PixelData
 from mmengine.model import BaseModel
 from torch import Tensor
 
-from mmseg.data import SegDataSample
 from mmseg.ops import resize
+from mmseg.structures import SegDataSample
 from mmseg.utils import (ForwardResults, OptConfigType, OptMultiConfig,
                          OptSampleList, SampleList)
