+++ mmseg/structures/seg_data_sample.py
