+++ tests/test_engine/test_visualization_hook.py
@@ -5,8 +5,8 @@ from unittest.mock import Mock
 import torch
 from mmengine.data import PixelData
 
-from mmseg.data import SegDataSample
 from mmseg.engine.hooks import SegVisualizationHook
+from mmseg.structures import SegDataSample
 from mmseg.visualization import SegLocalVisualizer
