+++ tests/test_sampler.py
@@ -2,8 +2,8 @@
 import pytest
 import torch
 
-from mmseg.data import OHEMPixelSampler
 from mmseg.models.decode_heads import FCNHead
+from mmseg.structures import OHEMPixelSampler
 
 
 def _context_for_ohem():
